feat(models): Comfy-Org 量化权重加载期反量化支持(int8_convrot/mxfp8/nvfp4)
数据治理并行进程遗留特性落地与处置： - 新增 quant_dequant.py：加载期将 Comfy-Org 发布的 int8_convrot(卷积旋转+哈达玛)、mxfp8、nvfp4 量化权重反量化为 fp16 再送推理，避免入库庞大 fp16 副本； - spec.precision_from_dit_model 解析含下划线尺寸(7b_sharp)与三种量化精度；config_models.ModelEntryConfig 新增对应 checkpoint/sha256/min_vram 字段； - seedvr2_engine/load_model 透传 precision，model_manager 精度感知加载与降级，common 路由透传； - restore.html 前端精度选择；scripts/download_model 新增 Comfy-Org 来源识别/解析/校验； - 许可：NOTICE 与 docs/LICENSE_COMPLIANCE.md 记录第三方量化权重许可与 sha256 溯源； - 测试：tests/test_quant_dequant.py 覆盖三种格式反量化契约与精度解析。 关联数据治理评估报告 P2-2(权重 sidecar/溯源) 与 P3-1(内容寻址血缘)。

+++ app/integrated_app/model_manager.py
@@ -273,7 +273,8 @@ class ModelManager:
             raise ValueError(f"未知的模型大小: {model_size}")
 
         if not self.check_model_exists(model_size, precision):
-            fallback_precision = "fp16" if precision == "fp8" else "fp8"
+            # 回退方向：fp16↔fp8 互备（numz 源）；Comfy-Org 量化精度缺失时回退 fp16
+            fallback_precision = "fp8" if precision == "fp16" else "fp16"
             if self.check_model_exists(model_size, fallback_precision):
                 logger.warning(f"{precision} 模型文件不存在，回退到 {fallback_precision}")
                 precision = fallback_precision
